+++ ui/plugin-system/src/components/PluginEditor/plugin-editor-api.ts
@@ -15,9 +15,9 @@ import { BoxProps } from '@mui/material';
 import { Definition, UnknownSpec, useEvent } from '@perses-dev/core';
 import { useState, useRef, useEffect } from 'react';
 import { produce } from 'immer';
-import { PluginType } from '../../model';
-import { PluginKindSelectProps } from '../PluginKindSelect';
-import { PluginSpecEditorProps } from '../PluginSpecEditor';
+import { PanelPlugin, PluginType } from '../../model';
+import { PluginKindSelectProps } from '../PluginKindSelect/PluginKindSelect';
+import { PluginSpecEditorProps } from '../PluginSpecEditor/PluginSpecEditor';
 import { usePlugin, usePluginRegistry } from '../../runtime';
 
 // Props on MUI Box that we don't want people to pass because we're either redefining them or providing them in
@@ -32,11 +32,14 @@ export interface PluginEditorProps extends Omit<BoxProps, OmittedMuiProps> {
 }
 
 type PreviousSpecState = Record<string, Record<string, UnknownSpec>>;
+type HideQueryEditorState = Record<string, boolean>;
 
 /**
  * Props needed by the usePluginEditor hook.
  */
-export type UsePluginEditorProps = Pick<PluginEditorProps, 'pluginType' | 'value' | 'onChange'>;
+export type UsePluginEditorProps = Pick<PluginEditorProps, 'pluginType' | 'value' | 'onChange'> & {
+  onHideQueryEditorChange?: (isHidden: boolean) => void;
+};
 
 /**
  * Returns the state/handlers that power the `PluginEditor` component. Useful for custom components that want to provide
@@ -45,16 +48,17 @@ export type UsePluginEditorProps = Pick<PluginEditorProps, 'pluginType' | 'value
  * kind back.
  */
 export function usePluginEditor(props: UsePluginEditorProps) {
-  const { pluginType, value } = props;
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  const { pluginType, value, onHideQueryEditorChange = () => {} } = props; // setting onHideQueryEditorChange to empty function here because useEvent requires a function
 
   // Keep a stable reference so we don't run the effect below when we don't need to
   const onChange = useEvent(props.onChange);
+  const onHideQuery = useEvent(onHideQueryEditorChange);
 
   // The previous spec state for PluginType and kind and a helper function for remembering current values
   const prevSpecState = useRef<PreviousSpecState>({
     [pluginType]: { [value.kind]: value.spec },
   });
-
   const rememberCurrentSpecState = useEvent(() => {
     let byPluginType = prevSpecState.current[pluginType];
     if (byPluginType === undefined) {
@@ -62,6 +66,11 @@ export function usePluginEditor(props: UsePluginEditorProps) {
       prevSpecState.current[pluginType] = byPluginType;
     }
     byPluginType[value.kind] = value.spec;
+  });
+
+  // The previous hide query state for each panel kind
+  const hideQueryState = useRef<HideQueryEditorState>({
+    [value.kind]: false,
   });
 
   const { defaultPluginKinds } = usePluginRegistry();
@@ -86,8 +95,17 @@ export function usePluginEditor(props: UsePluginEditorProps) {
       kind: pendingKind,
       spec: plugin.createInitialOptions(),
     });
+
+    if (pluginType === 'Panel') {
+      const panelPlugin = plugin as PanelPlugin;
+      hideQueryState.current[pendingKind] = !!panelPlugin.hideQueryEditor;
+      if (!!panelPlugin.hideQueryEditor !== hideQueryState.current[value.kind]) {
+        onHideQuery(!!panelPlugin.hideQueryEditor);
+      }
+    }
+
     setPendingKind('');
-  }, [pendingKind, plugin, rememberCurrentSpecState, onChange]);
+  }, [pendingKind, plugin, rememberCurrentSpecState, onChange, onHideQuery, hideQueryState, pluginType, value.kind]);
 
   /**
    * When the user tries to change the plugin kind, make sure we have the correct spec for that plugin kind before we
@@ -104,11 +122,18 @@ export function usePluginEditor(props: UsePluginEditorProps) {
         kind: nextKind,
         spec: previousState,
       });
-      return;
+    } else {
+      // Otherwise, kick off the async loading process
+      setPendingKind(nextKind);
     }
 
-    // Otherwise, kick off the async loading process
-    setPendingKind(nextKind);
+    if (
+      pluginType === 'Panel' &&
+      hideQueryState.current[nextKind] !== undefined &&
+      hideQueryState.current[value.kind] !== hideQueryState.current[nextKind]
+    ) {
+      onHideQuery(!!hideQueryState.current[nextKind]);
+    }
   };
 
   /**
@@ -122,5 +147,5 @@ export function usePluginEditor(props: UsePluginEditorProps) {
     );
   };
 
-  return { pendingKind, isLoading: isFetching, error, onKindChange, onSpecChange };
+  return { pendingKind, isLoading: isFetching, error, onKindChange, onSpecChange, rememberCurrentSpecState };
 }