+++ ui/plugin-system/src/runtime/index.ts
@@ -16,3 +16,4 @@ export * from './plugin-registry';
 export * from './template-variables';
 export * from './TimeRangeProvider';
 export * from './time-series-queries';
+export * from './DataQueriesProvider';